Landing Page
Created the landing page for my website

Correctly styled and componentized each piece of the first section of the design.

Followed a flex and grid design system. The website flows with a flex design and each section in the main component is built with a grid. 12 columns only atm since there is no responsive design implemented yet.

Added gsap animations for the start and the hello circle text. Currently only a spin animation.