Reformat landing page and styles
Modify the landing page to conform to a new design.

Added a landing image that contains my name and a blue star. This image is a svg which is embedded into the html so that I can access the blue star path to animate. The whole thing is a svg image instead of text so that it scales to the device with, with a max width of the layout sizing.

Changed the sub text to be 'creative front-end developer' instead of 'developer & designer'.

Removed the text circle that said 'Hello!'.