Updated design and flow
Updated the landing page to finalize the design. The background moves with the mouse, but snaps back to the center if the mouse moves outside of the viewport.

Header has been shrunk down a few sizes to fit with the design.

Works section is updated to fit with new design, although this is a work in progress.

The layout is a work-in-progress and needs to be simplified, perhaps to something with only the grid needed now?

Updated styles to fit with changes.
Updated animations to fit with changes.