Project routing
Created project routing so that users can view projects. This is automated through the use of the api and routing.

Created project data. This is stored in the data folder of the project and contains the information about the project.

Created project page to display the projects when the user clicks on the link to.

Created components for the api routing.
Created components for the page routing.

Header smile icon links back to home page.

Added a config file to tell the website to what route to use.

Modified styling and animations.
Formatted files.