Added banners that scroll
Added banners that scroll across the full width of the screen with large text. This required a re-do of the layout system that was being used to organize the pages. Moved to a single grid system where the overflow was constricted to a minimum width but had no max so it acted as free values.