Added footer and about preview
Added a footer to the website. This contains the contact information about me. Currently only links and no animations or effects added.

Added the image preview to the about me section so that when you hover over the about me title it shows a picture of me.

Updated styling, names, and animations.