Add mobile view
Add mobile view to the website.

Modified the header so that on the mobile view it displays a tray that will open up the menu instead of having the buttons on the top. Which is better for mobile.

Created a mixin to look for a mobile breakpoint width. This will then apply small changes to the website to have it fit  better on smaller screens. Like shrinking the text or setting full widths on elements.

Applied changes to the animations to make them work better on both desktop and mobile. Applied styling fixes.

Made the tray look better.