Small fixes
Applied small fixes that needed to be done to help polish the website.

Added noise to the website for an effect.

Made the 404 page all lowercase letters.

Made the page/head titles uppercase.

Fixed the animated line on initial entry where it appears for a split second and then goes to it's state.

Fixed the mobile tray where it would appear randomly or not appear in it's correct spot. Sometimes did not have the correct width or height, fixed.